backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
args==0.1.0
asyncpg==0.31.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.123.5
filelock==3.25.2
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifcopenshell==0.8.4
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.9
lark==1.3.1
lightgbm==4.6.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
openai==2.24.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prometheus_client==0.24.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.5
xgboost==3.2.0
